Fix camera control bug

+++ src/GraphicAPI/OpenGL/OpenGLTexture.hpp
@@ -36,7 +36,7 @@ public:
     inline GLenum textureType() const { return m_textureType; }
 
     #ifdef GFX_BUILD_IMGUI
-    inline void* imguiTextureId() const override { return (void*)(utils::uint64)textureID(); }
+    inline void* imguiTextureId() const override { return (void*)(uintptr_t)(utils::uint64)textureID(); }
     inline ImVec2 imguiUV0() const override { return ImVec2(0, 1); }
     inline ImVec2 imguiUV1() const override { return ImVec2(1, 0); }
     #endif
